working improvements, missing docs and mem tests

+++ TP/test-cases/make_test.sh
@@ -6,5 +6,7 @@ test_number=$1
 ../src/main < "in/$test_number.in" > tmp.txt
 diff tmp.txt "out/$test_number.out"
 ec=$?
-rm tmp.txt
+if [ $ec == 0 ]; then
+  rm tmp.txt
+fi
 exit $ec
